Changed from 3 components to one
